API-4759 : Code tidy up   - Added packages   - Extracted methods
