+++ src/main/scala/repos/GenerateGitCloneBashScript.scala
@@ -1,0 +1,15 @@
+package repos
+
+import uk.gov.hmrc.integrationcatalogueapiplatformtools.repos.CsvApiRecord
+
+object GenerateGitCloneBashScript {
+  def printScript(apisToProcess: Seq[CsvApiRecord]) = {
+
+    val text = apisToProcess
+      .map(api => s"git clone git@github.com:hmrc/${api.name}.git")
+      .mkString("\n")
+  
+    print(text)
+
+  }
+}
